Prune small MSA for Leah
Hey Leah, please put the alignment files that were generated during the
making of DB in the MembraneProteins folder. Set the matlab working
directory to MembraneProteins too and run this script. It will
calculate the number of sequences kept in the cluster analysis. I will
use that to keep only the good clusters.